Changed the timer to run every 6 hours, starting at midnight.
Updated some comments that were incorrect.

More comments for tracking where the script it up to in the logs.
Comments about what I need to do to make this even better on
the crypto side. Namely adding HMAC for file validation and
removing the temp directory where file are decrypted for
comparison. The less the data is in the clear, the safer
everyone is when using this script.